Add files via upload
In this project we will run Cohort analysis and compute metrics such as Retention rate and Churn rate. These metrics will allow us to understand customer trends. After that, we will compute Customer Lifetime Value (CLV) which is a measurement of how much a company expects to earn from an average customer in a lifetime. Then, we will run Recency/Frequency/Monetary/Tenure segmentation and create easy to interpret segments with K-mean clustering. On the top of that, we will predict customer transactions in the next month.